+++ docs/CHANGELOG.md
@@ -7,11 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Dat Files Models
-- Dat Files Routes
-- Treating of Dat Files
+- Dat Files Routes of updating and deletion
 - Dumping to Done Dat Files
 - Frontend
+- Changelog on different languages
+- GitHub Issues templates
+- GitHub Pull Requests templates
+- Contribution Guidelines
+
+## [0.0.2] - 2022-08-29
+
+### Created
+
+- Dat file model
+- Dat file Middleware
+- Dat file Controller
+- Interface for related `.dat` and `.done.dat` files
+- Trait for related `.dat` and `.done.dat` files
+
+### Implemented
+
+- List Dat API Route
+- Read Dat API Route
+- Store Dat API Route
+
+### Fixed
+
+- Incorrect code documentation styling
+- Incorrect spaces on some lines of code
+- Misspelled words
 
 ## [0.0.1] - 2022-08-17
 
@@ -41,5 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documented PHP files as of Laravel's Standards
 - Set documentation for Github's Community Standards
 
-[Unreleased]: <https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.1...HEAD>
+[Unreleased]: <https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.2...HEAD>
+[0.0.3]: <https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.2...v0.0.3>
+[0.0.2]: <https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.1...v0.0.2>
 [0.0.1]: <https://github.com/Mestre-Tramador/Exper-Dat-Reader/releases/tag/v0.0.1>